move exportDownload to annualupdate_minimal only
