+++ src/main/scala/org/globalforestwatch/summarystats/annualupdate_minimal/Adm2SummaryDF.scala
@@ -1,4 +1,4 @@
-package org.globalforestwatch.summarystats.annualupdate
+package org.globalforestwatch.summarystats.annualupdate_minimal
 
 import org.apache.spark.sql._
 import org.apache.spark.sql.functions._
