+++ src/main/scala/org/globalforestwatch/summarystats/annualupdate_minimal/AnnualUpdateMinimalExport.scala
@@ -2,7 +2,6 @@ package org.globalforestwatch.summarystats.annualupdate_minimal
 
 import org.apache.spark.sql.DataFrame
 import org.globalforestwatch.summarystats.SummaryExport
-import org.globalforestwatch.summarystats.annualupdate.AnnualUpdateExport
 
 object AnnualUpdateMinimalExport extends SummaryExport {
 
@@ -11,7 +10,7 @@ object AnnualUpdateMinimalExport extends SummaryExport {
                                     kwargs: Map[String, Any]): Unit = {
 
 
-    def exportArea(df: DataFrame): Unit = {
+    def exportSummary(df: DataFrame): Unit = {
 
       val adm2ApiDF = df.transform(Adm2ApiDF.sumArea)
       adm2ApiDF.write
@@ -47,17 +46,52 @@ object AnnualUpdateMinimalExport extends SummaryExport {
         .csv(path = outputUrl + "/iso/change")
     }
 
+    def exportDownload(df: DataFrame, outputUrl: String): Unit = {
+
+      val spark = df.sparkSession
+      import spark.implicits._
+
+      val adm2SummaryDF = df
+        .transform(Adm2SummaryDF.sumArea)
+
+      adm2SummaryDF
+        .transform(Adm2SummaryDF.roundValues)
+        .coalesce(1)
+        .orderBy($"country", $"subnational1", $"subnational2", $"treecover_density__threshold")
+        .write
+        .options(csvOptions)
+        .csv(path = outputUrl + "/adm2/download")
+
+      val adm1SummaryDF = adm2SummaryDF.transform(Adm1SummaryDF.sumArea)
+
+      adm1SummaryDF
+        .transform(Adm1SummaryDF.roundValues)
+        .coalesce(1)
+        .orderBy($"country", $"subnational1", $"treecover_density__threshold")
+        .write
+        .options(csvOptions)
+        .csv(path = outputUrl + "/adm1/download")
+
+      val isoSummaryDF = adm1SummaryDF.transform(IsoSummaryDF.sumArea)
+
+      isoSummaryDF
+        .transform(IsoSummaryDF.roundValues)
+        .coalesce(1)
+        .orderBy($"country", $"treecover_density__threshold")
+        .write
+        .options(csvOptions)
+        .csv(path = outputUrl + "/iso/download")
+
+    }
+
     val exportDF = summaryDF
       .transform(ApiDF.unpackValues)
-    //          apiDF.repartition($"iso")
-    //    val apiDF =
-    //    adm2DF
-    //      .transform(ApiDF.setNull)
+
     exportDF.cache()
 
-    exportArea(exportDF)
+    exportSummary(exportDF)
     exportChange(exportDF)
-    AnnualUpdateExport.exportSummary(exportDF, outputUrl)
+    exportDownload(exportDF, outputUrl)
 
     exportDF.unpersist()
